chore: improve structure
