+++ GomokuClient/src/App.tsx
@@ -38,47 +38,49 @@ const App = () => {
   };
 
   return (
-    <div className="font-open-sans flex flex-col items-center p-4 font-light">
-      <div className="mb-5 text-center">
-        {winner && (
-          <div className="mb-2 text-2xl">The Winner is: {winner}!</div>
-        )}
-      </div>
-      <div className="mb-5 flex w-full flex-wrap justify-center">
-        <div className="grid-cols-19 grid">
-          {board.map((row, rowIndex) => (
-            <div className="flex" key={rowIndex}>
-              {row.map((col, colIndex) => (
-                <Square
-                  key={`${rowIndex}-${colIndex}`}
-                  row={rowIndex}
-                  col={colIndex}
-                  value={col}
-                  onClick={handlePieceClick}
-                />
-              ))}
-            </div>
-          ))}
+    <div className="min-h-screen bg-[#161512] text-base text-[#bababa] sm:text-lg">
+      <div className="font-open-sans flex flex-col items-center p-4 font-light">
+        <div className="mb-5 text-center">
+          {winner && (
+            <div className="mb-2 text-2xl">The Winner is: {winner}!</div>
+          )}
         </div>
-        <div className="ml-4 flex flex-col justify-between">
-          <Card className="bg-white p-4">
-            <h2 className="mb-4 text-2xl font-semibold">Timer</h2>
-            <div className="mb-4 font-mono text-4xl">{formatTime(time)}</div>
-            <div className="flex flex-col gap-2">
-              <Button onClick={startTimer} disabled={isRunning}>
-                Start
-              </Button>
-              <Button onClick={stopTimer} disabled={!isRunning}>
-                Stop
-              </Button>
-              <Button onClick={resetTimer}>Reset</Button>
-              <Button className="mt-4" onClick={playAgain}>
-                Play again
-              </Button>
-            </div>
-            <br />
-            <Chat />
-          </Card>
+        <div className="mb-5 flex w-full flex-wrap justify-center">
+          <div className="grid-cols-19 grid">
+            {board.map((row, rowIndex) => (
+              <div className="flex" key={rowIndex}>
+                {row.map((col, colIndex) => (
+                  <Square
+                    key={`${rowIndex}-${colIndex}`}
+                    row={rowIndex}
+                    col={colIndex}
+                    value={col}
+                    onClick={handlePieceClick}
+                  />
+                ))}
+              </div>
+            ))}
+          </div>
+          <div className="ml-4 flex flex-col justify-between">
+            <Card className="bg-white p-4">
+              <h2 className="mb-4 text-2xl font-semibold">Timer</h2>
+              <div className="mb-4 font-mono text-4xl">{formatTime(time)}</div>
+              <div className="flex flex-col gap-2">
+                <Button onClick={startTimer} disabled={isRunning}>
+                  Start
+                </Button>
+                <Button onClick={stopTimer} disabled={!isRunning}>
+                  Stop
+                </Button>
+                <Button onClick={resetTimer}>Reset</Button>
+                <Button className="mt-4" onClick={playAgain}>
+                  Play again
+                </Button>
+              </div>
+              <br />
+              <Chat />
+            </Card>
+          </div>
         </div>
       </div>
     </div>
