+++ GomokuClient/src/components/FeaturedBoxes/index.ts
@@ -1,0 +1,1 @@
+export * from "./FeaturedBoxes.tsx";
