+++ GomokuClient/src/components/Tournamets/index.ts
@@ -1,0 +1,1 @@
+export * from "./Tournaments.tsx";
